refactor: unify corecrypto proteus impls into a common module
# Conflicts:
#	crypto-ffi/src/lib.rs
