+++ crypto-ffi/src/lib.rs
@@ -1,22 +1,3 @@
-#[macro_export]
-macro_rules! proteus_impl {
-    ($body:block or throw $err_type:ty) => {
-        {
-        cfg_if::cfg_if! {
-            if #[cfg(feature = "proteus")] {
-                #[allow(clippy::redundant_closure_call)]
-                $body
-            } else {
-                return <$err_type>::Err(core_crypto::Error::FeatureDisabled("proteus").into());
-            }
-        }
-        }
-    };
-    ($body:block) => {
-        proteus_impl!($body or throw ::std::result::Result<_, _>)
-    };
-}
-
 cfg_if::cfg_if! {
     if #[cfg(target_family = "wasm")] {
         mod wasm;
@@ -45,6 +26,7 @@ mod decrypted_message;
 mod error;
 mod identity;
 mod metadata;
+mod proteus;
 
 pub use bundles::{
     commit::CommitBundle, group_info::GroupInfoBundle, proposal::ProposalBundle,
